backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.68.0",
    "pydantic>=2.12.5",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "redis>=7.3.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.7",
    "types-PyMySQL>=1.1.0.20251103",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py312"
lint.select = [
    "F",  # pyflakes
    "E",  # pycodestyle
    "W",  # pycodestyle warnings
    "I",  # isort
    "D",  # pydocstyle
    "UP", # pyupgrade
    "N",  # pep8-naming
]
lint.ignore = [
    "D105", # undocumented-magic-method
    "D107", # undocumented-public-init
    "D205", # blank-line-after-summary
    "E402", # sys path before import
    "D415", # ends-in-punctuation
]
line-length = 100

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "D"] # __init__.pyは未使用インポートとドキュメントを無視
